Fix wrapText interface.
