+++ packages/flutter_tools/lib/src/commands/analyze_once.dart
@@ -132,7 +132,7 @@ class AnalyzeOnce extends AnalyzeBase {
       printStatus('');
     errors.sort();
     for (AnalysisError error in errors)
-      printStatus(wrapText(error.toString(), hangingIndent: 8));
+      printStatus(error.toString(), hangingIndent: 7);
 
     final String seconds = (timer.elapsedMilliseconds / 1000.0).toStringAsFixed(1);
